penambahan kolom stok minimal dan perbaikan cart

+++ database/migrations/2025_07_11_101902_create_cart_items_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id('cart_item_id');
             $table->unsignedBigInteger('cart_id');
             $table->unsignedBigInteger('product_id');
-            $table->integer('kuantitas');
+            $table->decimal('kuantitas');
             $table->integer('harga_satuan');
             $table->timestamps();
             $table->foreign('cart_id')->references('cart_id')->on('carts')->onDelete('cascade');
